Shift more stuff to `event.ts`.
